Fixed something that needs fixing

+++ lib/pages/progress/widgets/progress_bar_graph_logic.dart
@@ -62,7 +62,7 @@ class ProgressBarGraphLogic {
   // --- Computed Properties for the UI --- //
 
   double get totalCalories =>
-      filteredLogs.fold(0.0, (s, e) => s + e.kc.toDouble());
+      filteredLogs.fold(0.0, (s, e) => s + e.kc);
 
   double percentageOfTarget(double dailyGoal) {
     final weeklyTarget = dailyGoal * 7;
